Add mculib and libopencm3 as git submodules
and adapt Makefiles and the README file accordingly.

In the main Makefile, the LIBNAME variable is set before the include
statements because when set by the libopencm3/mk/genlink-config.mk
its value is empty at first, since it checks for the existence of the
library file, but this later is build by the $(OPENCM3_LIB) rule.
Kind of a chicken and egg problem.

The build instructions in the README file are now much simpler.